Dockerfile/requirements.txt
selenium==4.1.2
pytest==7.0.1
allure-pytest==2.9.45
pytest-xdist==2.5.0
